fix waring_as_error on CI Windows.

+++ src/plugins/intel_gpu/src/graph/impls/cm/paged_attention.cpp
@@ -157,8 +157,8 @@ public:
             auto out_shape = params.output_layouts[0].get_shape();
             const size_t kv_len = get_max_context_len(params) / STRIDE * STRIDE;
             const size_t q_len = out_shape[0];
-            const uint32_t M = q_len / STRIDE;   //# will slient drop the tails which is less than `stride`
-            const uint32_t N = kv_len / STRIDE;
+            const uint32_t M = static_cast<uint32_t>(q_len / STRIDE);   //# will slient drop the tails which is less than `stride`
+            const uint32_t N = static_cast<uint32_t>(kv_len / STRIDE);
             const size_t q_stride_pad = round_up_to(M, BLOCK_WG_M);
             const size_t N_kq_groups = ceil_div(N, BLOCK_WG_N);
 
@@ -169,8 +169,8 @@ public:
             if (block_size > 1) {
                 OPENVINO_ASSERT(block_size % STRIDE == 0, "sparse block_size must be devidable by stride.");
                 const uint32_t q_block_pad = ceil_div(q_len, block_size);
-                const uint32_t sum_per_token_in_block = block_size / STRIDE;
-                const uint32_t k_block_in_group = BLOCK_WG_N / sum_per_token_in_block;
+                const uint32_t sum_per_token_in_block = static_cast<uint32_t>(block_size / STRIDE);
+                const uint32_t k_block_in_group = static_cast<uint32_t>(BLOCK_WG_N / sum_per_token_in_block);
                 const uint32_t k_block_pad = k_block_in_group * N_kq_groups;
                 auto count_kq_exp_partial_sum = static_cast<int64_t>(desc->heads_num * q_stride_pad * k_block_pad);
                 internal_buffers.emplace_back(count_kq_exp_partial_sum, ov::element::f32);       // 3: kq_exp_partial_sum
